+++ src/adapter/ember/adapter/emberAdapter.ts
@@ -224,6 +224,7 @@ const WORKAROUND_JOIN_MANUF_IEEE_PREFIX_TO_CODE: {[ieeePrefix: string]: Zcl.Manu
     // NOTE: Lumi has a new prefix registered since 2021, in case they start using that one with new devices, it might need to be added here too...
     //       "0x18c23c" https://maclookup.app/vendors/lumi-united-technology-co-ltd
     '0x54ef44': Zcl.ManufacturerCode.LUMI_UNITED_TECHOLOGY_LTD_SHENZHEN,
+    '0x04cf8c': Zcl.ManufacturerCode.LUMI_UNITED_TECHOLOGY_LTD_SHENZHEN,
 };
 
 /**
@@ -402,6 +403,16 @@ export class EmberAdapter extends Adapter {
             case SLStatus.ZIGBEE_NETWORK_CLOSED: {
                 this.oneWaitress.resolveEvent(OneWaitressEvents.STACK_STATUS_NETWORK_CLOSED);
                 logger.info(`[STACK STATUS] Network closed.`, NS);
+
+                if (this.manufacturerCode !== DEFAULT_MANUFACTURER_CODE) {
+                    await this.queue.execute<void>(async () => {
+                        logger.debug(`[WORKAROUND] Reverting coordinator manufacturer code to default.`, NS);
+                        await this.ezsp.ezspSetManufacturerCode(DEFAULT_MANUFACTURER_CODE);
+
+                        this.manufacturerCode = DEFAULT_MANUFACTURER_CODE;
+                    });
+                }
+
                 break;
             }
             case SLStatus.ZIGBEE_CHANNEL_CHANGED: {
@@ -979,6 +990,7 @@ export class EmberAdapter extends Adapter {
                     Array.from(backup!.networkOptions.extendedPanId),
                     backup!.logicalChannel,
                     backup!.ezsp!.hashed_tclk!, // valid from getStoredBackup
+                    backup!.networkUpdateId,
                 );
 
                 result = 'restored';
@@ -995,6 +1007,7 @@ export class EmberAdapter extends Adapter {
                     this.networkOptions.extendedPanID!,
                     this.networkOptions.channelList[0],
                     randomBytes(EMBER_ENCRYPTION_KEY_SIZE), // rnd TC link key
+                    0,
                 );
 
                 result = 'reset';
@@ -1041,6 +1054,7 @@ export class EmberAdapter extends Adapter {
         extendedPanId: ExtendedPanId,
         radioChannel: number,
         tcLinkKey: Buffer,
+        nwkUpdateId: number,
     ): Promise<void> {
         const state: EmberInitialSecurityState = {
             bitmask:
@@ -1100,7 +1114,7 @@ export class EmberAdapter extends Adapter {
             radioChannel,
             joinMethod: EmberJoinMethod.MAC_ASSOCIATION,
             nwkManagerId: ZSpec.COORDINATOR_ADDRESS,
-            nwkUpdateId: 0,
+            nwkUpdateId,
             channels: ZSpec.ALL_802_15_4_CHANNELS_MASK,
         };
 
@@ -1678,16 +1692,17 @@ export class EmberAdapter extends Adapter {
                 panID,
                 extendedPanID: ZSpec.Utils.eui64LEBufferToHex(Buffer.from(extendedPanID)),
                 channel,
+                nwkUpdateID: this.networkCache.parameters.nwkUpdateId,
             };
         });
     }
 
     // queued
-    public async addInstallCode(ieeeAddress: string, key: Buffer): Promise<void> {
+    public async addInstallCode(ieeeAddress: string, key: Buffer, hashed: boolean): Promise<void> {
         return await this.queue.execute<void>(async () => {
             // Add the key to the transient key table.
             // This will be used while the DUT joins.
-            const impStatus = await this.ezsp.ezspImportTransientKey(ieeeAddress as EUI64, {contents: ZSpec.Utils.aes128MmoHash(key)});
+            const impStatus = await this.ezsp.ezspImportTransientKey(ieeeAddress as EUI64, {contents: hashed ? key : ZSpec.Utils.aes128MmoHash(key)});
 
             if (impStatus == SLStatus.OK) {
                 logger.debug(`[ADD INSTALL CODE] Success for '${ieeeAddress}'.`, NS);
@@ -1860,13 +1875,6 @@ export class EmberAdapter extends Adapter {
                     throw new Error(`[ZDO] Failed set join policy with status=${SLStatus[setJPstatus]}.`);
                 }
             } else {
-                if (this.manufacturerCode !== DEFAULT_MANUFACTURER_CODE) {
-                    logger.debug(`[WORKAROUND] Reverting coordinator manufacturer code to default.`, NS);
-                    await this.ezsp.ezspSetManufacturerCode(DEFAULT_MANUFACTURER_CODE);
-
-                    this.manufacturerCode = DEFAULT_MANUFACTURER_CODE;
-                }
-
                 await this.ezsp.ezspClearTransientLinkKeys();
 
                 const setJPstatus = await this.emberSetJoinPolicy(EmberJoinDecision.ALLOW_REJOINS_ONLY);